pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "waze_home"
version = "0.1.0"
description = "CLI tool to get the fastest route between home and work using Waze data"
requires-python = ">=3.10"
dependencies = [
    "click",
    "requests",
    "rich",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest]
testpaths = ["tests"]